+++ Shape/sphere.c
@@ -4,7 +4,7 @@
 
 
 
-t_point new_point(float x, float y, float z)
+t_point new_point(double x, double y, double z)
 {
 	t_point pos;
 
@@ -22,19 +22,19 @@ t_color add_color(const size_t *c)
 	obj.b = (int)c[2];
 	return obj;
 }
-bool 	intersectRaySphere(t_ray r, t_minirt *rt, float *t,t_hit **pHit)
+bool 	intersectRaySphere(t_ray r, t_minirt *rt, double *t,t_hit **pHit)
 {
 
-	float t_min;
+	double t_min;
 	t_point dist;
-	float discriminant;
-	float C;
-	float B;
-	float sqrt_discr;
-	float A;
+	double discriminant;
+	double C;
+	double B;
+	double sqrt_discr;
+	double A;
 	t_Sphere *closestSphere;
 	t_Sphere *s;
-	float hitDestance;
+	double hitDestance;
 	int i;
 	s = rt->Sphere;
 	 r.direction = normalizing(r.direction);
@@ -50,6 +50,7 @@ bool 	intersectRaySphere(t_ray r, t_minirt *rt, float *t,t_hit **pHit)
 		discriminant = B * B - 4.0f * A * C;
 		sqrt_discr = sqrtf(discriminant);
 		t_min = (-B - sqrt_discr) / (2.0f * A);
+//		printf("%f\n", t_min);
 		if (t_min < hitDestance)
 		{
 			hitDestance = t_min;
